Create sentiment_analysis.py
The project is on sentiment analysis. The project have used pre-trained word embeddings of 50 dimensions. Furthermore in the code, firstly  stop word  removing, tokenization and lemmatization is performed on the sentences, and at last padding is performed to make the all the sentences of a fixed length. Furthermore, in the code, LSTM neural network is designed as per the dataset and the classes of sentiment. For training we used a dummy dataset having reviews of hotels. Each review can be either positive or negative.